requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
files==1.1.1
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
